refactor: separate BlackTurnResult
