+++ src/test/kotlin/omok/model/GameStateTest.kt
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.Test
 class GameStateTest {
     @Test
     fun `흑돌의 턴에 오목을 만들면 흑돌의 승리를 반환한다`() {
-        val blackTurn: GameState.Playing = BlackTurn
+        val blackTurn: GameState.Playing = BlackTurn()
 
         val actual = blackTurn.play(BLACK_WIN_BOARD_IF_PUT_1_5, POSITION_1_5)
 
@@ -25,16 +25,16 @@ class GameStateTest {
 
     @Test
     fun `흑돌의 턴에 금수를 두면 흑돌의 턴을 반환한다`() {
-        val blackTurn: GameState.Playing = BlackTurn
+        val blackTurn: GameState.Playing = BlackTurn()
 
         val actual = blackTurn.play(BLACK_FORBIDDEN_BOARD_IF_PUT_1_1, POSITION_1_1)
 
-        assertThat(actual).isEqualTo(BlackTurn)
+        assertThat(actual).isEqualTo(BlackTurn())
     }
 
     @Test
     fun `흑돌의 턴이 끝날 때 오목도 금수도 아닌 경우 백돌의 턴을 반환한다`() {
-        val blackTurn: GameState.Playing = BlackTurn
+        val blackTurn: GameState.Playing = BlackTurn()
 
         val actual = blackTurn.play(EMPTY_BOARD, POSITION_1_1)
 
@@ -56,6 +56,6 @@ class GameStateTest {
 
         val actual = whiteTurn.play(EMPTY_BOARD, POSITION_1_0)
 
-        assertThat(actual).isEqualTo(BlackTurn)
+        assertThat(actual).isEqualTo(BlackTurn())
     }
 }
